feat(knowledge): index companion knowledge beside boards + nested diagram tree
Two related bodies of knowledge/diagram work bundled into one PR.

1) Companion knowledge indexing (this session's feature)
   The dark-sibling rule excluded EVERY non-board .md sitting in a board
   folder from index/recall/snapshot/dashboard — forcing a choice between
   good organization (board + detailed prose together) and recall-ability.
   Now role-based: a companion .md that declares itself as knowledge via
   `name:` frontmatter is indexed as first-class knowledge; generator
   scripts, spec JSON, and frontmatter-less helper notes stay dark.
   - isDarkDiagramSibling() gains an optional `isIndexableKnowledge` arg
     (defaults false → backward-safe); knowledge-index.ts and recall.ts
     read frontmatter once and pass the flag (no extra fs reads).
   - Co-located teardowns (e.g. acme/acme.excalidraw.md + acme.teardown.md)
     now recall normally without moving any file. Verified end-to-end via
     `knowledge index --plain` + `memory recall`.

2) Nested diagram tree + reorg (prior in-review knowledge-diagram-nesting)
   Dashboard renders knowledge/diagrams as a recursive nested folder tree
   (categories sub-group instead of collapsing flat); excalidraw cards show
   a sketch icon; 12 boards reorganized under system/, memory-science/,
   product/; build-all.mjs resolves boards by glob (path-aware); SKILL docs,
   prose refs, flat-regression + e2e tests updated.

Tests: full unit suite green (1979 passed). Docs updated in skill/SKILL.md
and skill-packs/excalidraw/SKILL.md.
