+++ scripts/diagrams/build-all.mjs
@@ -5,7 +5,7 @@
 //
 //   node scripts/diagrams/build-all.mjs
 import { execFileSync } from 'node:child_process';
-import { existsSync, mkdirSync } from 'node:fs';
+import { existsSync, mkdirSync, readdirSync } from 'node:fs';
 import { dirname, resolve } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { withRenderer } from './render-excalidraw.mjs';
@@ -13,6 +13,23 @@ import { withRenderer } from './render-excalidraw.mjs';
 const ROOT = resolve(dirname(fileURLToPath(import.meta.url)), '../..');
 const DIA = resolve(ROOT, '_dream_context/knowledge/diagrams');
 const OUT = resolve(ROOT, 'public/image');
+
+// Boards live in per-title folders, optionally grouped under a category subfolder
+// (e.g. diagrams/system/recall/recall.board.cjs). Resolve each board by basename
+// anywhere under DIA so the manifest stays a flat name → png map regardless of how
+// the boards are organized on disk.
+function findUnder(dir, name) {
+  for (const e of readdirSync(dir, { withFileTypes: true })) {
+    const full = resolve(dir, e.name);
+    if (e.isDirectory()) {
+      const hit = findUnder(full, name);
+      if (hit) return hit;
+    } else if (e.name === name) {
+      return full;
+    }
+  }
+  return null;
+}
 
 // name → { board spec (.cjs), rendered png, scale }. PNG basenames are the public
 // contract the docs reference; keep them stable.
@@ -32,8 +49,10 @@ if (!existsSync(OUT)) mkdirSync(OUT, { recursive: true });
 
 console.log('1/2  building boards (spec → .excalidraw.md)…');
 for (const d of DIAGRAMS) {
-  const boardPath = resolve(DIA, d.board);
-  if (!existsSync(boardPath)) { console.log(`  SKIP ${d.board} (not found)`); continue; }
+  const boardPath = findUnder(DIA, d.board);
+  if (!boardPath) { console.log(`  SKIP ${d.board} (not found)`); continue; }
+  // Generators write their .excalidraw.md next to themselves (__dirname), so the
+  // board folder's location is irrelevant to where the output lands.
   execFileSync('node', [boardPath], { stdio: 'pipe' });
   console.log(`  built ${d.board}`);
 }
@@ -41,8 +60,8 @@ for (const d of DIAGRAMS) {
 console.log('2/2  rendering PNGs (board → public/image)…');
 await withRenderer(async (render) => {
   for (const d of DIAGRAMS) {
-    const boardMd = resolve(DIA, d.board.replace('.board.cjs', '.excalidraw.md'));
-    if (!existsSync(boardMd)) continue;
+    const boardMd = findUnder(DIA, d.board.replace('.board.cjs', '.excalidraw.md'));
+    if (!boardMd) continue;
     await render(boardMd, resolve(OUT, d.png), d.scale);
     if (d.pdf) await render.pdf(boardMd, resolve(OUT, d.pdf), d.scale);
   }
